ISSUE #768 get rid of tests.js and move the test functions into the suite
